Added Topic 3 DQ 1 page for CSS history.
Assignment Instructions: Research the history of the CSS language. In a table, summarize the history of the language up to present time, with the features, advancements, and purpose of the language for each major point in time.
Instructor's additional request: Consider as you research this, how have browsers advanced the capabilities of CSS and the web in general?